pyproject.toml
[tool.poetry]
name = "paddleocr"
version = "0.0.1"  # You need to set a version
description = "Fork of PaddleOCR adapted for recognizing characters on German license plates."
authors = ["PaddlePaddle <Paddle-better@baidu.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/optocycle/PaddleOCR.git"
homepage = "https://github.com/optocycle/PaddleOCR"

[tool.poetry.dependencies]
python = ">=3.9" 
paddlepaddle-gpu = { url = "https://paddle-whl.bj.bcebos.com/stable/cu118/paddlepaddle-gpu/paddlepaddle_gpu-3.0.0rc1-cp310-cp310-linux_x86_64.whl" }
shapely = "*"
scikit-image = "*"
pyclipper = "*"
lmdb = "*"
tqdm = "*"
numpy = "*"
rapidfuzz = "*"
opencv-python = "*"
opencv-contrib-python = "*"
cython = "*"
Pillow = "*"
pyyaml = "*"
python-docx = "*"
beautifulsoup4 = "*"
fonttools = ">=4.24.0"
fire = ">=0.3.0"
requests = "*"
albumentations = "*"
albucore = "*"
setuptools = "^78.0.2"
mlflow = "^2.21.0"
python-dotenv = "^1.1.0"
paddle2onnx = "^2.0.0"
psutil = "^7.0.0"

[tool.poetry.scripts]
paddleocr = "paddleocr.paddleocr:main"